[ENHANCEMENT] Moved hidden file sourcing
When the init for hidden files was in the zsh_profile they would not be
available in new terminal tabs - which means the user has to restart the
terminal each time - very annoying.

This change will make it so the dev does not have to source the profile
all the time.