Add Improved VIM markdown support
This adds vim-markdown which provides some additional markdown support
and also adds in syntax highlighting improvements (obviously
subjective).

The additions to syntax are me hacking things together and there are
likely better way, but I am noobish at this point and really just want
to move on with my life.  My main goal with syntax configuration is to
make "special syntax" (bold, italics, inline code, links) all be
highlighted with the same color vs just the `**` of bold being
highlighted and the word itself being another color/style.  The purpose
of this is I use those to let my eye just dart around the content and
not have to read everything.