Update Nvim Config - Move to lua
The original motivation for this was because I wasn't sure if my vim
configurations for telescope which were embedded in my init.vim file
were working so I moved to lua, which is a good move in general, but
adds more to work with.  The long and the short is my configurations
were working, I was not understanding the tools.  So, here is what we
change in this commit:

- Move Telescope configuration to lua
- Update Telescope bindings to 1 letter chars
- Remove vimgrep_arguments
  - I was thinking this was making git ignored files visible, but this
    config is only used during live_grep and grep_search so we can
    remove because my issue was not with those

So mainly we cleanup our configuration and align our bindings to what we
want to use