nvim: set smarter search functionality
By defualy, search would be case sensitive.  With this change, it's case
insensitive unless I start using capitalization in the search. When I do
this, it becomes case sensitive.  I feel this will give me the best of
both worlds as I do like case sensitivity, but also want to opt out
without having to set it in nvim