enhance: init kitty
Decided to test kitty after hearing about how fast it is.  This commit
adds a kitty config to our dotfiles and the `kitty` to our PATH so that
we can use `kitty` in the terminal shell environment to open the app.

Aside from the above, the main thing we've done with the configs is add
keymappings for our nvim custom stuffs.  These were a little tricky to
wrap my head around and I don't know if my approach will have a negative
side effect, but here we are.